Add `PanicReason` to `ScriptResult`
The VM should not raise panic to the client from a valid transaction.

Instead, any erroneous state must generate a `ScriptResult::result` with
a panic description.

The introduced script result description will reserve its first 8-bits
to describe the error variant, and the remainder bits to add relevant
data to describe the error to the user.