Add mint transaction type. (#406)
Partially addresses #286.

Add a new transaction type, `Mint`, that only contains outputs. In the
general case, no validation of this transaction is necessary beyond
checking for well-formedness. Note that this PR doesn't actually allow
mint transactions anywhere, it only defines a new transaction type.

Two applications exist:
1. A coinbase tx, i.e. #286. This will require a block validation rule
to check that the first tx is a mint tx and that the values are
constrained.
2. Block-level flash loans.
https://twitter.com/aeyakovenko/status/1568429252332384257 A mint tx
must be paired with an equivalent burn (e.g. a tx with only inputs and
no outputs). Credits to @pluriholonomic for discussion on this.