Add data field to panic receipt (#409)
When the SDK is trying to estimate which contracts should be added to
the inputs of a transaction, it is only able to obtain a u8 panic reason
from the VM without any additional info about what contract ID was
expected to be in the inputs. While the VM knows what contract ID is
missing from the inputs, the ABI spec currently doesn't provide any way
for the VM to return it.

This PR adds an optional contractId field to panic receipts which enables
the VM to aid the access-list estimation process.