Specify block header (#405)
Fixes #284

Adds a specification for the block header. The header is split into two:
the application header (i.e. the Fuel transactions) and the consensus
header (i.e. the blockchain + consensus logic), which wraps the
application header. While not strictly necessary, this allows for a
cleaner separation of concerns logically. From an implementation
perspective, this separation doesn't actually need to be done today.

The intention is that a consensus signature for the current block will
be sent together with that block (header) to the base layer as a means
of authentication. This allows an untrusted relayer to relay signed
block headers to the base chain.