Chain Identifier (#462)
Closes #340

Rather than adding chain ID as a new field to the transaction, this
approach includes chain id as part of our identifier hashing procedures.
This means predicates, contracts and transactions will be innately chain
specific, preventing cross-chain replay.

Consider the scenario where there are two instances of fuel (e.g.
app-specific rollups) that each have the same contracts deployed. If a
user signs a transaction to transfer tokens without a chain identifier,
the same transaction could potentially be replayed on the second chain
unintentionally. While this would require the same input utxos to exist
on both chains, there are still certain scenarios where this could occur
and cause a single transaction to be used for multiple transfers across
different chains.

I considered hashing `CHAIN_ID`'s into Message ID's, but didn't see any
obvious benefits since it wouldn't be emitted by or committed to on L1.
It may be worth considering an alternative approach such as a new
chain_id field for messages to ensure the same message can't be used on
multiple chains. However, since predicate messages are primarily used in
the current bridging designs, this should already be covered for the
majority of use-cases.

---------

Co-authored-by: ControlCplusControlV <44706811+ControlCplusControlV@users.noreply.github.com>