increase code root merkle leaf size to 16KiB (#476)
Calculating the merkle root of across every instruction pair is turning
out to be quite cost prohibitive, as noted in this issue:
https://github.com/FuelLabs/fuel-vm/issues/385

Increasing the leaf size when computing a code root drastically reduces
the sha256 overhead. I assume the original design was chosen because 8
bytes is the standard word size of the EVM, making it easier to pass
into a merkle proof function.

Increasing the leaf size leads to >30x performance improvement on the
execution layer. Fraud challenges are crypto-economically guaranteed to
happen infrequently. So it seems worthwhile to increase the fraud proof
cost in the worst case in exchange for substantially improving the happy
path of regular transaction execution.

I also made a benchmarking framework to experiment with various leaf
sizes, making it easy to verify the improvement due to leaf size
increases: https://github.com/Voxelot/merkle-bench