docs: Add Merkle tree specs to cryptographic primitives (#504)
Related issues:
- https://github.com/FuelLabs/fuel-vm/issues/501#issue-1792237720

This PR gives the Fuel specs its own Merkle tree specification, and
removes links to the archived Celestia spec. This gives Fuel ownership
over its own Merkle tree specification and allows us to adapt it to our
design.

Specifically, this PR copies the specification for the binary and sparse
Merkle trees as described in the Celestia spec. Additional comments are
my own, and describe in more detail topics like the BMT proof format.

Lastly, this PR removes the specification for the binary Merkle sum tree
(BMST). The BMST is currently not supported by Fuel and does not fall
under the Fuel specification at this time.