Transaction Policies (#514)
# Summary

Use a dynamic set of policies for transaction validation requirements
such as gaslimit & maturity. This will reduce the size of transactions
that always leave these fields unset, and also future proof the
transaction format for later requirements such as multi-dimensional
pricing.

A new policy that's been added is `witnessLimit`. Having a witness limit
allows the network to charge for witness data without exposing users to
malicious basefee manipulation by third-parties / relayers. Without a
charge & limit for witness data, blockspace could be easily saturated
with spurious spammy witness data.

Another significant change is to separate base costs of the transaction
from the transaction gas limit. Gas limit will now only apply to script
execution, and everything else is charged as part of the minimum
transaction fee.

Since the gas limit can't restrict the total cost of the transaction
anymore, another policy called `MaxFee` was added. This is a more direct
way to place an upper bound on transaction costs and is also future
compatible with multi-dimensional pricing.

## Multi-dimension pricing preparation

In order for resources such as execution and storage to be charged for
independently, transactions must be able to specify an anticipated
consumption limit for each resource kind. By abstracting these limits
into a dynamic set of policies, new resource types can be added in the
future more easily.

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>