Added a new `Upgrade` transaction to perform a network upgrade (#567)
Corresponding implementation:
https://github.com/FuelLabs/fuel-vm/pull/707

The change adds a new `Upgrade` transaction that allows upgrading either
consensus parameters or state transition function used by the network to
produce future blocks.

The purpose of the upgrade is defined by the `Upgrade Purpose` type.

The `Upgrade` transaction is chargeable, and the sender should pay for
it. Transaction inputs should contain only base assets.

Only the privileged address can upgrade the network. The privileged
address can be either a real account or a predicate.

We use postcard algorithm to serialize and deserialize consensus
parameters during the upgrade. This algorithm works with numbers in a
more sufficient way and compresses them well.