Clarify that non-zero reserved part of an instruction causes vm panic (#583)
VM PR: https://github.com/FuelLabs/fuel-vm/pull/737

What is says on the tin. This is a small but breaking change. It's
unlikely to cause any breakage, as no Sway code or code constructed
using the fuel-asm helper functions is affected.

This change is originating from an audit report that pointed out the
likely unintentional behavior of the unused part of instruction is
ignored.

### Before requesting review
- [x] I have reviewed the code myself

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>