Blob tx and instructions (#592)
Work towards https://github.com/FuelLabs/fuel-specs/issues/589. The VM
PR https://github.com/FuelLabs/fuel-vm/pull/780 contains a working
implementation.

To allow contracts larger than the contract size limit, and to allow
upgradability, we'll be adding a new tx type that's just a bunch of
bytes, and opcodes to load subsections of that. A *blob* will be like a
contract, but without state or balances, allowing it to be quite a bit
cheaper. The new instructions `BSIZ` and `BLDD` and the new blob mode
for `LDC` closely follow how `CSIZ`, `LDC` and `CCP` work, but only
operation on blob data.

### Before requesting review
- [x] I have reviewed the code myself