Remove erroneous "Notes" from `SUB` and `SUBI` instructions (#613)
### Description

This PR fixes the documentation of the `SUB` and `SUBI` instructions and
makes it follow the same pattern as the `ADD` and `ADDI` instructions.

The documentation was wrongly using the term "overflow" in the "Notes"
of those two instructions, instead of "underflow" (as used in the longer
explanation below the instruction).

The PR simply removes the erroneous description from the "Notes" and
leaves them empty, as in `ADD` and `ADDI`.

### Before requesting review
- [x] I have reviewed the code myself