Add SUB operation to NIOP (#637)
When implemention `niop`, I discovered that the subtraction via `sub`
instruction doesn't work for narrower integer types, as the result will
not be properly truncated if the operation overflows. This PR adds
subtraction mode, corresponding the the commit
https://github.com/FuelLabs/fuel-vm/pull/922/commits/99433f9889112de59d231a44ad954c9a534e376e
in the fuel-vm branch. SInce the `niop` isn't implemented yet, we can
freely change the operation table to keep it more consistent.


### Before requesting review
- [x] I have reviewed the code myself